Fixed an issue with the 6502 irq emulation where the cycles for the next instruction were calculated, the machine was emulated for that many cycles, then the 6502 instruction was executed.
The problem with this is that if during the cycles for that instruction an interrupt was raised, the actual instruction executed would be the first one if the irq. In real hw, thats like an irq travelling back in time and causing the CPU to execute a different instruction. It also has the side effect of the wrong number of cycles being executed of the rest of the machine for that instruction.

Now, the cycles are calculated for the next cpu instruction, and the machine is emulated for that many cycles as before, but now, when the CPU instruction is executed, it always executes the instruction used to calculate the cycles. This then behaves like the real hw; an irq only happens once the current instruction has finished.